+++ lib/src/controllers/drawables/shape/arrow_drawable.dart
@@ -21,25 +21,15 @@ class ArrowDrawable extends Sized1DDrawable implements ShapeDrawable {
   ArrowDrawable({
     Paint? paint,
     this.arrowHeadSize,
-    required double length,
-    required Offset position,
-    double rotationAngle = 0,
-    double scale = 1,
-    Set<ObjectDrawableAssist> assists = const <ObjectDrawableAssist>{},
-    Map<ObjectDrawableAssist, Paint> assistPaints =
-        const <ObjectDrawableAssist, Paint>{},
-    bool locked = false,
-    bool hidden = false,
-  })  : paint = paint ?? ShapeDrawable.defaultPaint,
-        super(
-            length: length,
-            position: position,
-            rotationAngle: rotationAngle,
-            scale: scale,
-            assists: assists,
-            assistPaints: assistPaints,
-            locked: locked,
-            hidden: hidden);
+    required super.length,
+    required super.position,
+    super.rotationAngle,
+    super.scale,
+    super.assists,
+    super.assistPaints,
+    super.locked,
+    super.hidden,
+  }) : paint = paint ?? ShapeDrawable.defaultPaint;
 
   /// The actual arrow head size used in drawing.
   double get _arrowHeadSize => arrowHeadSize ?? paint.strokeWidth * 3;