+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 ## [1.1.0]
 
 * Update internal dependencies
-* Update minimum Dart version to 3.0
-* Update minimum Flutter version to 3.10
+* Update minimum Dart version to 2.0
+* Update minimum Flutter version to 3.16
 
 ## [1.0.1]
 * Fix a bug where `TextDrawable`s would incorrectly render on the final image ([Issue #19](https://github.com/omarhurani/flutter_painter/issues/19)).